% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpval.R
\name{bootpval}
\alias{bootpval}
\title{The Bootstrap P-Value Function}
\usage{
bootpval(x, conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")
}
\arguments{
\item{x}{sample from the population, used for one-sample t.test for p-value}

\item{conf.level}{confidence level, 1 - alpha, for the confidence interval for mu}

\item{iter}{used for iterations of bootstrap intervals, increase for more accuracy}

\item{mu0}{NULL value to test sample mean against for rejecting or not rejecting the NULL hypothesis H_0}

\item{test}{test type, "two" for two-tailed test, "upper" for right-tailed test, or "lower" for left-tailed test}
}
\value{
named list with p-value, t_calc, length of vector, data vector passed, test type passed, and confidence interval
}
\description{
The Bootstrap P-Value Function
}
\examples{
bootpval(x = rnorm(30, 25, 5), mu0 = 24)
}
