% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{The Mean Confidence Interval Function}
\usage{
myci(x, alpha = 0.05, rnd = 4)
}
\arguments{
\item{x}{sample from the population to find mean of}

\item{alpha}{probability outside the confidence interval}

\item{rnd}{number of decimal places to round to in output}
}
\value{
(1-alpha)100% confidence interval for population mean in vector of length 2
}
\description{
The Mean Confidence Interval Function
}
\examples{
myci(x = rnorm(30, 10, 12), alpha = 0.05)
}
