% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{The CLT Normal from Uniform Function}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{size of each sample}

\item{iter}{number of samples}

\item{a}{lower limit of uniform}

\item{b}{upper limit of uniform}
}
\value{
the vector of sample means
}
\description{
The CLT Normal from Uniform Function
}
\examples{
mycltu(n = 20, iter = 10000, a = 0, b = 10)
}
