% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{The Log Likelihood Maximization Function}
\usage{
mymaxlikg(lfun, theta)
}
\arguments{
\item{lfun}{log likelihood function of theta, commonly represented as l(theta)}

\item{theta}{sequence of theta values}
}
\value{
theta estimate to maximize log likelihood and hence likelihood
}
\description{
The Log Likelihood Maximization Function
}
\examples{
mymaxlikg(lfun=function(p){log(dbinom(2, 6, p)*dbinom(4, 10, p))}, theta=seq(0, 1, length = 10000))
}
