% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynbin.R
\name{mynbin}
\alias{mynbin}
\title{The Negative Binomial Probability Function}
\usage{
mynbin(y, r, p)
}
\arguments{
\item{y}{number of trials until rth success}

\item{r}{number of successes to evaluate on}

\item{p}{probability of success in each trial}
}
\value{
probability of y trials for rth success when probability of success is p
}
\description{
The Negative Binomial Probability Function
}
\examples{
mynbin(10,3,0.4)
}
