% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{The CSV Reading Function}
\usage{
myread(csv)
}
\arguments{
\item{csv}{a csv file path}
}
\value{
a data frame containing a representation of the data in the file
}
\description{
The CSV Reading Function
}
\examples{
myread("MTBE.csv")
}
