% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{The Optimal Plane Tickets Function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{amount of seats on plane}

\item{gamma}{intended probability of overbooking}

\item{p}{individual probability of a show}
}
\value{
a named list with optimal tickets, solved with a discrete distribution (nd) and solved with a continuous distribution (nc), and inputs N, p, and gamma
}
\description{
The Optimal Plane Tickets Function
}
\examples{
ntickets(N = 400, gamma = 0.02, p = 0.95)
}
